Refactor logging module and update imports to use `_logging`
